// apps/public-app/src/app/core/models/event-type.enum.ts
export enum EventType {
  RENTAL_REQUEST_CREATED = "RENTAL_REQUEST_CREATED",
  RENTAL_REQUEST_ACCEPTED = "RENTAL_REQUEST_ACCEPTED",
  RENTAL_REQUEST_REJECTED = "RENTAL_REQUEST_REJECTED",
  PAYMENT_RECEIVED = "PAYMENT_RECEIVED",
  CONTRACT_CREATED = "CONTRACT_CREATED",
  KEY_DELIVERED = "KEY_DELIVERED"
}